add s21_fabs() with tests

+++ include/s21_math.h
@@ -22,7 +22,7 @@ long double s21_atan(double x);
 long double s21_ceil(double x);
 long double s21_cos(double x);
 long double s21_exp(double x);
-long double s21_fabs(double x);
+double s21_fabs(double x);
 long double s21_floor(double x);
 long double s21_fmod(double x, double y);
 long double s21_log(double x);
